Added messasge commands
